cleared warnings

+++ LearnOpenGL/Sources/Mesh.cpp
@@ -33,7 +33,7 @@ void Mesh::recomputePerVertexNormals (bool angleBased) {
 
 	const unsigned int n = m_vertexPositions.size();
 	std::vector<std::list<glm::uvec3>> neighbours(n);
-	for (glm::vec3 triangle : m_triangleIndices) {
+	for (glm::uvec3 triangle : m_triangleIndices) {
 		neighbours[triangle.x].push_back(triangle);
 		neighbours[triangle.y].push_back(triangle);
 		neighbours[triangle.z].push_back(triangle);
